+++ src/main/java/service/factories/SenderServiceSingleton.java
@@ -1,21 +1,25 @@
 package service.factories;
 
-import service.SenderService;
-import service.api.ISenderService;
+import dao.factories.DAOType;
+import dao.factories.EmailSendingDAOSingleton;
+import service.EmailSendingService;
+import service.api.ISendingService;
 
 public class SenderServiceSingleton {
 
-    private volatile static ISenderService instance = null;
+    private volatile static ISendingService instance = null;
 
     private SenderServiceSingleton() {
     }
 
-    public static ISenderService getInstance() {
+    public static ISendingService getInstance() {
         if (instance == null) {
             synchronized (SenderServiceSingleton.class) {
                 if (instance == null) {
-                    instance = new SenderService(GenreServiceSingleton.getInstance(),
-                            ArtistServiceSingleton.getInstance());
+                    instance = new EmailSendingService(
+                            GenreServiceSingleton.getInstance(),
+                            ArtistServiceSingleton.getInstance(),
+                            EmailSendingDAOSingleton.getInstance(DAOType.DB));
                 }
             }
         }
